Fix `xy_color` annotation

+++ zha/application/platforms/light/__init__.py
@@ -225,7 +225,7 @@ class BaseLight(BaseEntity, ABC):
         effect: str | None = None,
         flash: FlashMode | None = None,
         color_temp: int | None = None,
-        xy_color: tuple[int, int] | None = None,
+        xy_color: tuple[float, float] | None = None,
     ) -> None:
         """Turn the entity on."""
 
@@ -321,7 +321,7 @@ class BaseClusterHandlerLight(BaseLight):
         effect: str | None = None,
         flash: FlashMode | None = None,
         color_temp: int | None = None,
-        xy_color: tuple[int, int] | None = None,
+        xy_color: tuple[float, float] | None = None,
     ) -> None:
         """Turn the entity on."""
         duration = (
@@ -402,7 +402,7 @@ class BaseClusterHandlerLight(BaseLight):
         effect: str | None,
         flash: FlashMode | None,
         color_temp: int | None,
-        xy_color: tuple[int, int] | None,
+        xy_color: tuple[float, float] | None,
         duration: float,
         execute_if_off_supported: bool,
         brightness_supported: bool,
@@ -1404,7 +1404,7 @@ class LightGroup(BaseClusterHandlerLight, GroupEntity):
         effect: str | None = None,
         flash: FlashMode | None = None,
         color_temp: int | None = None,
-        xy_color: tuple[int, int] | None = None,
+        xy_color: tuple[float, float] | None = None,
     ) -> None:
         """Send an assume event to all members of the group."""
